requirements.txt
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
asgiref==3.10.0
azure-communication-email==1.1.0
azure-core==1.37.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
drf-yasg==1.21.11
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-genai==1.60.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
inflection==0.5.1
isodate==0.7.2
jiter==0.12.0
packaging==25.0
pillow==12.0.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.2
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
resend==2.19.0
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.11.0
